Move inlineFailCases to Rewrite.hs
