+++ liquidhaskell-boot/liquidhaskell-boot.cabal
@@ -82,7 +82,6 @@ library
                       Language.Haskell.Liquid.Termination.Structural
                       Language.Haskell.Liquid.Transforms.ANF
                       Language.Haskell.Liquid.Transforms.CoreToLogic
-                      Language.Haskell.Liquid.Transforms.Rec
                       Language.Haskell.Liquid.Transforms.RefSplit
                       Language.Haskell.Liquid.Transforms.Rewrite
                       Language.Haskell.Liquid.Transforms.Simplify
